(zlists) mm: Accommodate CDM nodes in MPOL_BIND implementation
Currently MPOL_MBIND interface simply fails on a coherent device memory
node after the zonelist changes introduced earlier. Without __GFP_THISNODE
flag, the first node of the nodemask will not be selected in the case where
the local node (where the application is executing) is not part of the user
provided nodemask for MPOL_MBIND. This will be the case for coherent memory
nodes which are always CPU less. This patch changes the mbind() system call
implementation so that memory can be allocated from coherent memory node
through MPOL_MBIND interface.

Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>